+++ sphinx/source/changelog.rst
@@ -11,6 +11,7 @@ Changelog
 -----
 * Added the :ref:`tap2sna.py` command (for building snapshots from TAP/TZX
   files)
+* Added support to :ref:`skool2html.py` for multiple CSS themes
 * Moved the ``Font`` and ``StyleSheet`` parameters from the ``[Paths]`` section
   to the :ref:`ref-game` section
 * Moved the ``JavaScript`` parameter from the ``[Paths]`` section to the